\docType{package}
\name{fastqUtils}
\alias{fastqUtils}
\alias{fastqUtils-package}
\title{fastqUtils: Utilities for demultiplexing illumina fastqs and running dada2 workflow.}
\description{
The package \pkg{fastqUtils} provides functions for demultiplexing illumina amplicon libraries and running the dada2 workflow.
}
\details{
The package \pkg{fastqUtils} depends on \pkg{doMC}, \pkg{seqinr}, \pkg{ShortRead}, \pkg{Biostrings}, \pkg{dada2}, \pkg{ggplot2}, which can be installed from CRAN or Bioconductor.

To see the preferable citation of the package, type \code{citation("fastqUtils")}.
}
\author{
Tristan Cordier <trco@norceresearch.no>
}
